+++ src/test/testData/state/outsideFactory/essential/caretAtArrayKeyInDirectlyPassedShortClosure.php
@@ -1,0 +1,26 @@
+<?php
+
+class Id {
+    public string $value;
+}
+
+class Account {
+    public function __construct(
+        public readonly string $id,
+        public readonly int $age,
+    ) {}
+}
+
+class AccountFactory extends ClassFactory {
+    protected string $class = Account::class;
+
+    protected function definition(): array
+    {
+        return [
+            'id' => new Id(),
+            'age' => 1,
+        ];
+    }
+}
+
+AccountFactory::new()->state(fn (array $attributes) => ['<caret>' => $attributes['id']]);